test: rename rule no_repetitive_steps to no_repetitive_step_keyword
