+++ lib/config/gherlintrc.js
@@ -19,6 +19,6 @@ module.exports = {
      */
     rules: {
         indentation: ["warn", 2],
-        no_repetitive_steps: "error",
+        no_repetitive_step_keyword: "error",
     },
 };
