+++ lib/rules/Rule.js
@@ -15,7 +15,7 @@ module.exports = class Rule {
         return this._problems;
     }
 
-    addProblem(problem) {
+    storeLintProblem(problem) {
         this._problems.push(problem);
     }
 };
