+++ lib/rules/index.js
@@ -3,6 +3,6 @@ module.exports = {
     ruleOptions: require("./_ruleOptions"),
     Rules: {
         indentation: require("./indentation"),
-        no_repetitive_steps: require("./no_repetitive_steps"),
+        no_repetitive_step_keyword: require("./no_repetitive_step_keyword"),
     },
 };
