+++ lib/rules/no_repetitive_step_keyword.js
@@ -7,8 +7,8 @@ module.exports = class NoRepetitiveSteps extends Rule {
      * Rule filename and RuleId MUST follow snake_case convention
      */
     static meta = {
-        ruleId: "no_repetitive_steps",
-        message: "Repetitive steps '%s'",
+        ruleId: "no_repetitive_step_keyword",
+        message: "Repetitive step keyword '%s'",
         type: "error",
         location: {},
         hasFix: true,
@@ -60,7 +60,7 @@ module.exports = class NoRepetitiveSteps extends Rule {
                         keyword.trim()
                     ),
                 };
-                this.addProblem(problem);
+                this.storeLintProblem(problem);
             }
             // 'And' and '*' can be used repeatedly
             if (keyword.trim() === "And" || keyword.trim() === "*") return;
